Help when running in Docker Desktop (#180)
Running the Akita agent in Docker Desktop (macOS or Windows) will fail to
capture traffic to services running on the host.

Rather than sharing the host networking stack, Docker Desktop runs the daemon
in a VM.  Docker users are instructed to use host.internal.docker to
communicate with the host.

We can use the presence of that DNS name to detect if we're running in a
container with Docker Desktop.  If so, suggest using the native macOS binary if
the agent fails to capture traffic.