[POA-1340] Add a mechanism to randomly _not_ start apidump. (#17)
Controlled by a flag and a new environment variable.

The use case here is that we want to run the agent but some, but not
all, of the pods in a critical deployment. This is not a great solution,
but it's much easier than the alternatives considered. The deployment
can include a probability as an environment variable that any given
apidump command will run; those that don't run will just hang forever to
avoid any problems caused by container exit.

One downside is that changing the environment variable causes all the
pods to be restarted.