[POA-1514] Preserve POSTMAN_ENV in envFile in ec2 start (#27)
If a user uses the `ec2 start` command with POSTMAN_ENV env var, its
value is not set in the env file, which stores other env vars for the
systemd service. This will cause an issue for the apidump command since
it will then default to prod env.

In this PR, we are preserving the POSTMAN_ENV value in the env file and
using that in apidump command which is running as a systemd service.