[POA-2129] Separate out common `apidump` flags (#49)
* In this PR, we have moved out common `apidump` flags from `apidump.go`
to `common_flags.go`.
* `common_flags.go` have two functions:
* `AddCommonApiDumpFlags()`: Add the flags to cobra.Command object. It
returns the objects to which flag values will be assigned.
* `ConvertCommonApiDumpFlagsToArgs()`: Returns an array of strings (like
`--flag value`). It is added to the apidump command in different
platform installation configurations.
* The `ecs` and `ec2` commands will read these values and then convert
them to flags for `apidump` commands and add them to the definition file
(`ecs`) or systems service file (`ec2`)
* The `apidump` command will use these flag values to set the main
`apidump` package variables.