[POA-2711] Implementation of CRI functions (#69)
In this PR, we have implemented CRI APIs modules for interaction with
the container runtime

Function Implemented:
* NewCRIClient: Initialize the runtimeService client, opens a GRPC
connection to the CRI endpoint
* GetNetworkNamespace: Fetch network namespace for given container ID
* GetEnvVars: Fetch env vars for given container ID

Some extra files:
* model.go: It stores the model objects. There is not defined model from
K8s as of now so it could change in future.
* util.go: It consists of 2 main functions: 1) parse the endpoint and 2)
Convert the inspect container response to golang object
* remote_runtime.go: It consists of functions which create a connection
with the container runtime and initialize the object.

Note: Most of the functions are referenced from
[cri-client](https://github.com/kubernetes/cri-client) repository of k8s