[POA-3003] Handle deployment tags for daemonSet's apidump processes (#86)
Changes in this PR:

* Changes in the deployment package's functions to accept the envVars
map as the input
* Use these functions while creating podArgs object in the kube run
command
* Add check in the apidump process to not call deployment.UpdateTags()
function again if the process is started by the parent `kube run`
command

**Note:** Here, for `hostname`, we expect the `HOSTNAME` environment
variable to be set, but it is required. It is set for our pods in
Postman, but it is not set in a pod running in Minikube locally.