[POA-2849] Add systemd restart for already running service in ec2 (#89)
This pull request includes changes to the `cmd/internal/ec2/add.go` file
to enhance the handling of the `postman-insights-agent` service.

Enhancements to service handling:

*
[`cmd/internal/ec2/add.go`](diffhunk://#diff-de4c0d4aa9b0bc2881e1a3420db0ab595602fc4d945fbad5383e1586cd62a8d5R36-R38):
Added a new constant `active` to represent the active state of the
`postman-insights-agent` service.
*
[`cmd/internal/ec2/add.go`](diffhunk://#diff-de4c0d4aa9b0bc2881e1a3420db0ab595602fc4d945fbad5383e1586cd62a8d5R275-R294):
Modified the `enablePostmanInsightsAgent` function to check if the
`postman-insights-agent` service is active, and if so, restart it with
the new configurations. This includes handling errors and providing
informative messages to the user.