Remove get services calls from active commands (#114)
In this PR, we have removed the usage of the `GetServices()` function in
the front client because this endpoint uses a deprecated endpoint, which
will be removed soon and will not work with Postman authentication.

Changes:
* Updated the specs command to take serviceID input as a `--project`
flag.
* Also updated `GetTraceURIByTags` function to take serviceID as a
parameter instead of looking for it from BE.
* I have to not removed the `GetServices()` function yet since it is
still being used by some deprecated commands, like nginx which will be
removed as a part of agent code clean-up.