Add a supervisor process to the agent (#112)
With this patch, we use the agent process to be it's own supervisor.
This is controlled with __X_AKITA_NO_FORK environment variable. If it's
set to "1", the process won't fork itself and we will have the old
behaviour. For all other values, the first thing it is going to do is
fork and execute itself. The root process will then monitor the child
process and restart it if necessary.