[POA-3499] Remove separate package for interface mock (#121)
In this PR, we have moved the `interface.go` mock file outside of its
own package to the main `rest` package.

**Why?**
Since we don't push mock files to VCS _(which is the right practice)_
the `mock` package is missing from the publicly available package. This
was causing issues with the `go mod tidy` command in the superstar
service, where the package is used.
```
go: finding module for package github.com/postmanlabs/postman-insights-agent/rest/mock
go: akitasoftware.com/superstar/cli-postman imports
        github.com/postmanlabs/postman-insights-agent/cmd imports
        github.com/postmanlabs/postman-insights-agent/pcap imports
        github.com/postmanlabs/postman-insights-agent/trace tested by
        github.com/postmanlabs/postman-insights-agent/trace.test imports
        github.com/postmanlabs/postman-insights-agent/rest/mock: module github.com/postmanlabs/postman-insights-agent@latest found (v0.37.0), but does not contain package github.com/postmanlabs/postman-insights-agent/rest/mock
```

**Behaviour after changes:**
* After moving the `mock_interface.go` file to the parent `rest`
package, we are not getting a missing import error while running `go mod
tidy`, though the functions will be missing in the
`backend_collection_test.go` file, but it should be okay since this file
will never be used by the superstar service.

**Other:**
Also done small linting and typo fixes.