[POA-3089] Change telemetry package to handle daemon set sub processes (#139)
In this PR, we have changed the telemetry package to provide telemetry
support for subprocesses (go routines) under the main processes.

Changes:
- Telemetry package changes:
  - Move the global functions under an interface and struct model.
- Created a root (Default) telemetry object, which is used by the
existing processes for sending telemetry
  - A new constructor which will create a new scoped telemetry object.
- Removed the irrelevant analytics functions which were used in
deprecated functions.
- apidump and rest client changes:
- Change the way the error report function is set in the base client,
not it will be passed as a parameter to the constructor, just like the
auth handler
- Set a global package telemetry variable in the apidump package to use
it in various functions instead of passing it as a parameter down the
chain.
- Changes in pcap and trace package functions to use the telemetry
function used by the apidump process. _(Yes, this is contradicting to
what we did in the previous point)_
- Other syntactical changes and rearrangements.