[POA-4440] Add capture all payload support in kube run command (#145)
In this PR, we have added support for setting paths for which we will
always capture payloads in the `kube run` command.
The user will set this as an env var at their pod level, where the value
will be a JSONified string.

Changes:
* Add a new env var which will capture the string value
* Convert this string value to slices of string by trying to convert it
to a JSON object and default to an empty string. It errors